requirements.txt
fastapi
uvicorn[standard]
pyjwt
erppeek
requests
PyYAML>=6.0
line-bot-sdk
passlib
gspread
oauth2client